+++ localization_common/include/localization_common/cloud_registration/icp_registration.hpp
@@ -30,8 +30,8 @@ public:
   IcpRegistration(float max_corr_dist, float trans_eps, float euc_fitness_eps, int max_iter);
 
   bool set_target(const PointCloudPtr & target) override;
-  bool match(const PointCloudPtr & input, const Eigen::Matrix4f & initial_pose) override;
-  Eigen::Matrix4f get_final_pose() override;
+  bool match(const PointCloudPtr & input, const Eigen::Matrix4d & initial_pose) override;
+  Eigen::Matrix4d get_final_pose() override;
   double get_fitness_score() override;
   void print_info() override;
