+++ localization_common/include/localization_common/cloud_registration/icp_svd_registration.hpp
@@ -33,8 +33,8 @@ public:
   IcpSvdRegistration(float max_corr_dist, float trans_eps, float euc_fitness_eps, int max_iter);
 
   bool set_target(const PointCloudPtr & target) override;
-  bool match(const PointCloudPtr & input, const Eigen::Matrix4f & initial_pose) override;
-  Eigen::Matrix4f get_final_pose() override;
+  bool match(const PointCloudPtr & input, const Eigen::Matrix4d & initial_pose) override;
+  Eigen::Matrix4d get_final_pose() override;
   double get_fitness_score() override;
   void print_info() override;
 
@@ -59,7 +59,7 @@ private:
   PointCloudPtr input_source_;
 
   Eigen::Matrix4f transformation_;
-  Eigen::Matrix4f final_pose_;
+  Eigen::Matrix4d final_pose_;
 };
 
 }  // namespace localization_common
